Implemented Gradle
This was hard
