+++ app/src/main/java/UML/DiagramController.java
@@ -1,3 +1,5 @@
+package UML;
+
 import java.util.ArrayList;
 
 public class DiagramController
